chore: refactor to the qa service filename
