+++ sage/chat.py
@@ -1,8 +1,8 @@
 # chat.py
 import chainlit as cl
 
+from sage.sources.qa import SourceQAService
 from sage.utils.ai_tools import load_duck_search, load_jira_tools, load_qa_tool
-from sage.utils.source_qa import SourceQAService
 
 sage_tools = load_duck_search() + load_jira_tools() + load_qa_tool()
 qa_chat = SourceQAService(mode="chat", tools=sage_tools)
