+++ sage/sources/utils.py
@@ -8,6 +8,7 @@ from chainlit.user_session import UserSession
 from langchain.memory import ConversationBufferWindowMemory
 from langchain.schema.document import Document
 from langchain.schema.runnable import RunnableLambda
+
 from sage.constants import SENTINEL_PATH, logger
 from sage.utils.sources import Source
